Extract the shared Negotiate authenticator into kerberos_common.go
